feat: use macro in i16
