+++ core/src/data/value/binary_op/integer/mod.rs
@@ -1,3 +1,4 @@
+mod i16;
 mod i8;
 mod u8;
